pyproject.toml
[project]
name = "fus"
version = "0.1.0"
description = ""
authors = [
    {name = "LtttAZ",email = "louis_zhang_1303@126.com"}
]
requires-python = "==3.13.2"
dependencies = [
    "typer (>=0.23.1,<0.24.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "platformdirs (>=4.9.1,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "azure-devops (>=7.1.0,<8.0.0)"
]


[tool.poetry]
packages = [
    {include = "cli", from = "src"},
    {include = "common", from = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
